+++ app/src/main/java/zwaggerboyz/instaswaggify/filters/GaussianBlurFilter.java
@@ -25,8 +25,8 @@ public class GaussianBlurFilter extends AbstractFilterClass {
         };
 
         /* slider default value */
-        mValues = new int[] {
-                0
+        mValues = new float[] {
+                0f
         };
     }
